[!!!][FEATURE] Latest Posts Node Type
This change introduces a new node type for displaying a list of posts
sorted by publication date. It can either be used as a teaser element
for showing one or a couple of posts, or for a page which displays a
lists of all available posts with endless scrolling.

Note that this commit removes the plugin which previously displayed
the list of posts.

Known issues: Directly after inserting a Latest Posts node you'll see
a TypoScript error. Just reload the page to solve this.